# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quantum Commerce API
  description: API oficial do case fictício Quantum Commerce — disciplina Tools, Automations and Workflows (FIAP). Backend
    determinístico para agentes n8n, Dify, Power Automate e Python. Sem LLM no backend.
  version: 2.0.0
paths:
...
  /api/v1/inventory/{sku}:
    get:
      tags:
      - Inventory
      summary: Get Inventory Route
      operationId: get_inventory
      parameters:
      - name: sku
        in: path
        required: true
        schema:
          type: string
          title: Sku
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/inventory/{sku}/availability:
    get:
      tags:
      - Inventory
      summary: Get Inventory Availability Route
      operationId: get_inventory_availability
      parameters:
      - name: sku
        in: path
        required: true
        schema:
          type: string
          title: Sku
      - name: postal_code
        in: query
        required: true
        schema:
          type: string
          title: Postal Code
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/interactions:
    post:
      tags:
      - Interactions
      summary: Create Interaction Route
      operationId: create_interaction
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InteractionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Interactions
      summary: List Interactions Route
      operationId: list_interactions
      parameters:
      - name: customer_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Customer Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InteractionDetail'
                title: Response List Interactions
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/interactions/{interaction_id}:
    get:
      tags:
      - Interactions
      summary: Get Interaction Route
      operationId: get_interaction
      parameters:
      - name: interaction_id
        in: path
        required: true
        schema:
          type: string
          title: Interaction Id
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InteractionDetail'
        '404':
          description: Interaction not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/lab/slow:
    get:
      tags:
      - Lab
      summary: Lab Slow
      operationId: lab_slow
      parameters:
      - name: seconds
        in: query
        required: false
        schema:
          type: integer
          maximum: 15
          minimum: 1
          default: 5
          title: Seconds
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Slow
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lab/error:
    get:
      tags:
      - Lab
      summary: Lab Error
      operationId: lab_error
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Error
        '500':
          description: Simulated internal server error
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lab/rate-limit:
    get:
      tags:
      - Lab
      summary: Lab Rate Limit
      operationId: lab_rate_limit
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Rate Limit
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lab/not-found:
    get:
      tags:
      - Lab
      summary: Lab Not Found
      operationId: lab_not_found
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Not Found
        '404':
          description: Resource not found (simulated)
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/lab/validation:
    post:
      tags:
      - Lab
      summary: Lab Validation
      operationId: lab_validation
      parameters:
      - name: X-Lab-Group
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Lab-Group
      - name: X-Request-ID
        in: header
        required: false
        schema:
          type: string
          default: anonymous
          title: X-Request-Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationPayload'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Lab Validation
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    AvailabilityResponse:
      properties:
        sku:
          type: string
          title: Sku
        available:
          type: boolean
          title: Available
        quantity:
          type: integer
          title: Quantity
        estimated_delivery_days:
          type: integer
          title: Estimated Delivery Days
      type: object
      required:
      - sku
      - available
      - quantity
      - estimated_delivery_days
      title: AvailabilityResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InteractionCreate:
      properties:
        customer_id:
          type: string
          title: Customer Id
        channel:
          type: string
          title: Channel
        message:
          type: string
          title: Message
        response:
          type: string
          title: Response
        source:
          type: string
          title: Source
          default: ai-agent
      type: object
      required:
      - customer_id
      - channel
      - message
      - response
      title: InteractionCreate
    InteractionDetail:
      properties:
        interaction_id:
          type: string
          title: Interaction Id
        customer_id:
          type: string
          title: Customer Id
        channel:
          type: string
          title: Channel
        message:
          type: string
          title: Message
        response:
          type: string
          title: Response
        source:
          type: string
          title: Source
        lab_group:
          type: string
          title: Lab Group
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - interaction_id
      - customer_id
      - channel
      - message
      - response
      - source
      - lab_group
      - created_at
      title: InteractionDetail
    InteractionResponse:
      properties:
        interaction_id:
          type: string
          title: Interaction Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - interaction_id
      - created_at
      title: InteractionResponse
    InventoryResponse:
      properties:
        sku:
          type: string
          title: Sku
        total_available:
          type: integer
          title: Total Available
        warehouses:
          items:
            $ref: '#/components/schemas/WarehouseStock'
          type: array
          title: Warehouses
      type: object
      required:
      - sku
      - total_available
      - warehouses
      title: InventoryResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ValidationPayload:
      properties:
        email:
          type: string
          format: email
          title: Email
        amount:
          type: number
          minimum: 0.0
          title: Amount
      type: object
      required:
      - email
      - amount
      title: ValidationPayload
      examples:
      - amount: 42.5
        email: buyer@quantum.example
    WarehouseStock:
      properties:
        warehouse_id:
          type: string
          title: Warehouse Id
        city:
          type: string
          title: City
        available:
          type: integer
          title: Available
        reserved:
          type: integer
          title: Reserved
      type: object
      required:
      - warehouse_id
      - city
      - available
      - reserved
      title: WarehouseStock